cleanup: input text test
