+++ tests/unit/input-text-widget.spec.ts
@@ -122,4 +122,67 @@ describe('Widget Input Text', () => {
     const wrapper = mount(InputTextWidget);
     expect(wrapper.find('.field__msg-warning').exists()).toBeFalsy();
   });
+
+  it('should instantiate', () => {
+    const wrapper = shallowMount(InputTextWidget);
+
+    expect(wrapper.exists()).toBeTruthy();
+  });
+
+  it('should have a label', () => {
+    const wrapper = shallowMount(InputTextWidget, {
+      propsData: {
+        name: 'Stark',
+      },
+    });
+    const labelWrapper = wrapper.find('label');
+    expect(labelWrapper.text()).toEqual('Stark');
+  });
+
+  it('should have an empty placeholder', () => {
+    const wrapper = shallowMount(InputTextWidget, {
+      propsData: {
+        value: 'foo',
+      },
+    });
+    const el = wrapper.find("input[type='text']").element as HTMLInputElement;
+    expect(el.placeholder).toEqual('');
+  });
+
+  it('should have a placeholder', () => {
+    const wrapper = shallowMount(InputTextWidget, {
+      propsData: {
+        placeholder: 'I m a placeholder',
+        value: 'foo',
+      },
+    });
+    const el = wrapper.find("input[type='text']").element as HTMLInputElement;
+    expect(el.placeholder).toEqual('I m a placeholder');
+  });
+
+  it('should have an empty input', () => {
+    const wrapper = shallowMount(InputTextWidget);
+    const el = wrapper.find("input[type='text']").element as HTMLInputElement;
+    expect(el.value).toEqual('');
+  });
+
+  it('should have a non empty input', () => {
+    const wrapper = shallowMount(InputTextWidget, {
+      propsData: { value: 'foo' },
+    });
+    const el = wrapper.find("input[type='text']").element as HTMLInputElement;
+    expect(el.value).toEqual('foo');
+  });
+
+  it('should emit "input" event on update', () => {
+    const wrapper = shallowMount(InputTextWidget, {
+      propsData: {
+        value: ['Foo'],
+      },
+    });
+    const inputWrapper = wrapper.find('input[type="text"]');
+    (inputWrapper.element as HTMLInputElement).value = 'Bar';
+    inputWrapper.trigger('input', { value: 'Bar' });
+    expect(wrapper.emitted()).toEqual({ input: [['Bar']] });
+  });
 });
